fix: switch to focusing on Azure OpenAI
