+++ extensions/azure-openai-chat/.eslintrc.json
