+++ extensions/azure-openai-chat/.gitignore
