+++ extensions/azure-openai-chat/.prettierrc
