+++ extensions/azure-openai-chat/LICENSE
