+++ extensions/azure-openai-chat/README.md
@@ -1,0 +1,11 @@
+# Azure OpenAI Chat
+
+An Azure OpenAI client for Raycast.
+
+![client interface](/metadata/openai-chat-1.png)
+
+## Features
+
+1. Only one entry point required.
+2. Quickly rewrite prompts.
+3. Can archive conversations and reload them.
