+++ extensions/azure-openai-chat/assets/command-icon.png
