+++ extensions/azure-openai-chat/metadata/openai-chat-2.png
