+++ extensions/azure-openai-chat/src/components/MainView.tsx
