+++ extensions/azure-openai-chat/src/components/index.ts
