+++ extensions/azure-openai-chat/src/context/IndexContext.ts
