+++ extensions/azure-openai-chat/src/context/index.ts
