+++ extensions/azure-openai-chat/src/hooks/index.ts
@@ -1,3 +1,3 @@
 export * from "./useArchive";
 export * from "./useHistories";
-export * from "./useAI";
+export * from "./useAzureOpenAI";
