+++ extensions/azure-openai-chat/src/utils/formatContent.ts
