+++ extensions/azure-openai-chat/src/utils/generateMassages.ts
