+++ extensions/azure-openai-chat/src/utils/index.ts
@@ -1,4 +1,3 @@
-export * from "./openAI";
 export * from "./azureOpenAI";
 export * from "./formatContent";
 export * from "./generateMassages";
