--- extensions/openai-chat/README.md
@@ -1,19 +1,0 @@
-# OpenAI Chat
-
-An OpenAI client that supports OpenAI and Azure OpenAI.
-
-![client interface](/metadata/openai-chat-1.png)
-
-## Features
-
-1. Only one entry point required.
-2. Quickly rewrite prompts.
-3. Supports OpenAI and Azure OpenAI.
-4. Can archive conversations and reload them.
-
-## Preferences
-
-Endpoint: `https://api.openai.com` by default. You can input an Azure endpoint if you choose to use Azure OpenAI.  
-API Key: Input the API Key for your OpenAI.  
-Model: Please provide the name of the OpenAI-supported model or enter the deployment name of your Azure OpenAI.
-System Prompt: Setup the prompt of system
